+++ java-io-base/src/main/java/internal/io/EncodingFileFormatter.java
@@ -1,17 +1,20 @@
 package internal.io;
 
 import lombok.NonNull;
+import nbbrd.design.DecoratorPattern;
 import nbbrd.io.FileFormatter;
+import nbbrd.io.function.IOFunction;
 import nbbrd.io.function.IOSupplier;
-import nbbrd.io.function.IOUnaryOperator;
 
 import java.io.File;
+import java.io.FilterOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Path;
 
 import static nbbrd.io.Resource.uncloseableOutputStream;
 
+@DecoratorPattern
 @lombok.RequiredArgsConstructor
 public final class EncodingFileFormatter<T> implements FileFormatter<T> {
 
@@ -19,7 +22,7 @@ public final class EncodingFileFormatter<T> implements FileFormatter<T> {
     final FileFormatter<T> formatter;
 
     @NonNull
-    final IOUnaryOperator<OutputStream> encoder;
+    final IOFunction<OutputStream, ? extends FilterOutputStream> encoder;
 
     @Override
     public void formatFile(@NonNull T value, @NonNull File target) throws IOException {