+++ java-io-base/src/test/java/_test/io/ForwardingInputStream.java
@@ -1,23 +1,24 @@
 /*
  * Copyright 2018 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package _test.io;
 
 import nbbrd.io.function.IORunnable;
 
+import java.io.FilterInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -25,11 +26,11 @@ import java.io.InputStream;
  *
  * @author Philippe Charles
  */
-@lombok.AllArgsConstructor
-public class ForwardingInputStream extends InputStream {
+public class ForwardingInputStream extends FilterInputStream {
 
-    @lombok.experimental.Delegate
-    private final InputStream delegate;
+    public ForwardingInputStream(InputStream delegate) {
+        super(delegate);
+    }
 
     public ForwardingInputStream onClose(IORunnable onClose) {
         return new ForwardingInputStream(this) {