+++ CHANGELOG.md
@@ -7,6 +7,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.34] - 2025-05-08
+
+### Added
+
+- Add collector-based static factory to TextParser
+- Add more process shortcuts to TextParser
+
+### Changed
+
+- Replace gzip factories with filter factories in `File*` utilities
+
 ## [0.0.33] - 2025-03-06
 
 ### Added
@@ -292,7 +303,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.33...HEAD
+[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.34...HEAD
+[0.0.34]: https://github.com/nbbrd/java-io-util/compare/v0.0.33...v0.0.34
 [0.0.33]: https://github.com/nbbrd/java-io-util/compare/v0.0.32...v0.0.33
 [0.0.32]: https://github.com/nbbrd/java-io-util/compare/v0.0.31...v0.0.32
 [0.0.31]: https://github.com/nbbrd/java-io-util/compare/v0.0.30...v0.0.31